course compute servers -- code refactoring
